+++ Bybit.Exchange.Net/Bybit.Exchange.Net/Models/V5/Trade/GetOrderHistory.cs
@@ -140,7 +140,7 @@ namespace Bybit.Exchange.Net.Models.V5.Trade
         /// </item>
         /// </list>
         /// </summary>
-        public int? startTime { get; set; } = default!;
+        public double? startTime { get; set; } = default!;
 
         /// <summary>
         /// <value>Property <c>endTime</c></value>
@@ -148,7 +148,7 @@ namespace Bybit.Exchange.Net.Models.V5.Trade
         /// The end timestamp (ms)
         /// </para>
         /// </summary>
-        public int? endTime { get; set; } = default!;
+        public double? endTime { get; set; } = default!;
 
         /// <summary>
         /// <value>Property <c>limit</c></value>
