+++ Bybit.Exchange.Net/Bybit.Exchange.Net/Models/V5/Trade/GetTradeHistory.cs
@@ -76,7 +76,7 @@ namespace Bybit.Exchange.Net.Models.V5.Trade
         /// If both are passed, the rule is endTime - startTime &lt;= 7 days
         /// </para>
         /// </summary>
-        public int? startTime { get; set; } = default!;
+        public double? startTime { get; set; } = default!;
 
         /// <summary>
         /// <value>Property <c>endTime</c></value>
@@ -84,7 +84,7 @@ namespace Bybit.Exchange.Net.Models.V5.Trade
         /// The end timestamp (ms)
         /// </para>
         /// </summary>
-        public int? endTime { get; set; } = default!;
+        public double? endTime { get; set; } = default!;
 
         /// <summary>
         /// <value>Property <c>execType</c></value>
